"tbdiff" implementation for series-diff
This patch adds an implementation of "git-tbdiff" using a Javascript
implementation of the Hungarian algorithm for assignment.  The sorting
logic for the patches is the same as in tbdiff.

On the Python side, the patches are not concatenated anymore; they are
passed as (subject, body) tuples and each is placed in a separate textarea.