cli: clone recursively only for tester
Both tester and apply (applier mode) subcommands call git_clone_repo.
"--recursive" option was added so that repos that contain submodules
will likely just work with all the submodules present (although that can
be optimized too, to save time and bandwidth), but it is not necessary
for appliers and can significantly slow down the process because the
submodule repos are not cached.

Signed-off-by: Fam Zheng <famz@redhat.com>
Message-Id: <20180301025034.20913-1-famz@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>