cut number of queries for series detail view
The series detail views perform a lot of project queries to retrieve
m.project.name.  The project however is always known in advance; passing
it directly to prepare_message as an argument is not the cleanest thing,
but using prefetch_related/select_related does not fit Patchew's code
very well: find_series_head already calls first() to retrieve a Message,
and it is also ugly to sneak an annotate(project_name=F('project__name'))
or select_related('project') in find_series_head.  The full solution
to avoid adding parameters like this one would probably involve class-based
views, but for now this already cuts one query for each patch in the current
series.
Message-Id: <20180502155616.3826-3-pbonzini@redhat.com>

Signed-off-by: Fam Zheng <famz@redhat.com>