rest: POST for message endpoint
Creation of a message is handled by the new "create" method in
MessageManager, which also takes care of calling save_mbox().  The code
currently has some code duplication with add_message_from_mbox(), which
will be fixed soon.  The "create" name follows Django idioms more
closely and is what DRF expects.

However, due to the nested AddressSerializer, the CreateModelMixin is
not enough, we have to define a "create" method in BaseMessageSerializer
too.

Message-Id: <20180508115051.60844-3-shubhamjain7495@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>