mbox: text to JSON converter
MboxMessage already contains the parser for mbox files; add a method to
produce the same JSON representation used by the REST API, which will be
the basis for a text/plain parser in api.rest.

A little bit of code can be shared between the AddressSerializer and the
converter.  To avoid duplication, put it in a function addr_db_to_rest.
An alternative could be to move AddressSerializer in mbox.

Message-Id: <20180508193207.7663-2-shubhamjain7495@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>