rest: add api/v1/messages endpoint (POST support)
The endpoint applies a single message to the project according to the user
that is sending the request and the groups it belongs. If the user is an
importer, the message is added to all recognised projects. For other
users, the message is added to all the projects that are recognised and
maintained by the user.

Message-Id: <20180515120228.47478-1-shubhamjain7495@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>