rest: testcases and fixes around authorization
The tests check for:
- user that is not a maintainer of any project should not result in any
message being imported
- user that is not a maintainer of a project, but is in the importer
groups, should result in the message being imported to all recognized
projects
- user that is a maintainer of a project and is not in the importer
group, should result in the message being imported to maintained and
recognized projects

Also, changed the permission level of the MessageView to allow all
authenticated users. An authenticated user that is neither a maintainer
nor an importer would be able to send a POST request but not to create
a message, because the message would not be imported to any project.

Message-Id: <20180518023658.6008-1-shubhamjain7495@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>