search: split handling of anything but +/-/! to a separate function, simplify !
Define "+" as "imply an is: if there is no colon" and "-" as "imply a not:
if there is no colon".  It is possible however to write "-has:replies"
for example.

For simplicity, only allow one of +/-/!.  It is not possible anymore to
write "!!x" to mean "x".

The advantage is that _make_filter can just return the Q object directly.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>