Generalize Review model to QueuedSeries
The concept of a queue is a per-user model for maintainers. A maintainer
can have one ore more queues. The current Review model already has
accept and reject stautus, so we generalize that to support a more
flexible queue feature.

Previously, "mark as accepted" adds a (user, message, accept=True)
record in the Review model. Now it adds a (user, message, name='accept')
record in the QueuedSeries model.

Similary, for 'mark as rejected', a (user, message, name='reject')
record is added in the QueuedSeries model.

This makes the two queue names specially purposed on the web interface
and search terms, which we'll extend to also support normal queues.

Signed-off-by: Fam Zheng <famz@redhat.com>