views: fix Content-Transfer-Encoding and Content-Type of downloaded mbox
set_payload expects the argument to be already encoded according to the
message's content type and transfer encoding.  Force a good charset
and transfer encoding, in case the source message's is incompatible
with the trailers.  Otherwise, if the source message specified for
example an iso-8859-1 encoding, we get mojibake in the utf-8 trailers.

Otherwise, adding an 8-bit trailer to a 7-bit message causes the email
module to switch to quoted-printable, without the body actually being
quoted-printable.  As a result, all equals sign in the original message
are broken.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>