models: avoid IntegrityError when clearing log property
Result's log property provides some magic to hide the existence of
the LogEntry table (which in turn is only there to optimize access
to results, since the logs are rarely needed).

However, deleting the entry directly in the setter does not work,
because at this point the entry is still referenced in the Result
table.  Postgres complains about a foreign key constraint violation
when the setter calls entry.delete().

To fix this, which shows up as an error in the git-reset action,
detect at save time whether the log_entry became None, and if so
delete it _after_ saving the model.  While I am touching the code,
fix the override of save() to correctly pass the arguments to the
superclass.