Apply PEP8 to tests modules
The most of the changes in this commit are related with indentation and
whitespaces. However it was also necessary to fix how the import from
patchewtest members was implemented.

Before this change each test module was appending to sys.path the tests
directory so it was possible to import the base class PatchewTestCase.
However we can accomplish the same without change the sys.path
attribute. It's just a matter of turn tests directory into a Python
package (adding __init__.py file into it) and use the import statement
with relative imports as following:

    `from .patchewtest import PatchewTestcase`

The changes were validated running the tests with the following command:
`./manage.py test tests`.

Signed-off-by: Caio Carrara <ccarrara@redhat.com>
Message-Id: <20181128215617.9284-3-ccarrara@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>