patchew-importer: number import invocation in the logs
When importing a large number of messages it can be useful to know how the process
is progressing.  Offlineimap numbers the message it imports, but the subsequent
import phase does not.  Add a number to the import, which can be correlated with
the logs that were printed during the offlineimap phase.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>