applier: synchronize tags lazily into the cache
Every time the patchew's applier mode is restarted, it repopulates a cache
of git repositories.  For QEMU, the git repo has thousands of tags
which take a very long time to create.  Synchronize refs/heads/* only
(--mirror=fetch would synchronize all of refs/*); the tags are
only used by patchew-cli when a Based-on trailer is found, and can
be populated lazily.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>